refactor: extract WorkspaceState chat list into an extension file
Move the chat-list subscription/listener, delta application, row enrichment
and metadata resolution, peer-profile/sender resolution, and chat sorting
into WorkspaceState+ChatList.swift. Also promote the #if DEBUG sender-fetch
instrumentation counter from private(set) to internal (the 2.0 relaxation
only covered top-level-indented members, not ones nested in a #if block),
since the moved messageSenderProfiles mutates it.

Pure code movement, no behavior change. Build green; all 208 tests pass.
